+++ docs/changelog.rst
@@ -10,6 +10,8 @@ Changelog
 * Add horizontal line charts (thanks @chartique #301)
 * There is now a `formatter` config option to format values as specified. The formatter callable may or may not take `chart`, `serie` and `index` as argument. The default value formatting is now chart dependent and is value_formatter for most graph but could be a combination of value_formatter and x_value_formatter for dual charts.
 * The `human_readable` option has been removed. Now you have to use the pygal.formatters.human_readable formatter (value_formatter=human_readable instead of human_readable=True)
+* New chart type: SolidGauge (thanks @chartique #295)
+* Fix range option for some Charts (#297 #298)
 
 2.1.1
 =====